Use int64 for metrics instrument values
The float64 API inherited from the Prometheus client truncated
fractional values silently and had undefined behavior for NaN/Inf.
All trabbits metrics are integral counts, so accept and return int64
directly, including the health MetricsReporter interface.

+++ proxy.go
@@ -327,7 +327,7 @@ func (p *Proxy) sortNodesByLeastConnections(nodes []string) []string {
 
 	var nodeInfos []nodeInfo
 	for _, addr := range nodes {
-		connections := int64(p.metrics.UpstreamConnections.WithLabelValues(addr).Value())
+		connections := p.metrics.UpstreamConnections.WithLabelValues(addr).Value()
 		nodeInfos = append(nodeInfos, nodeInfo{addr: addr, connections: connections})
 	}
